refactor(compiler-sfc): move related files into separate directories
