+++ packages/compiler-sfc/src/style/pluginTrim.ts
