+++ packages/compiler-sfc/src/template/templateUtils.ts
