+++ packages/compiler-sfc/src/template/transformAssetUrl.ts
